# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Motor Ping API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/orders": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "Listar",
        "operationId": "listar_orders_get",
        "parameters": [
          {
            "name": "open",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Open"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100000,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "Crear",
        "operationId": "crear_orders_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrdenEntrada"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{orden_id}": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "Obtener",
        "operationId": "obtener_orders__orden_id__get",
        "parameters": [
          {
            "name": "orden_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Orden Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "orders"
        ],
        "summary": "Editar",
        "operationId": "editar_orders__orden_id__patch",
        "parameters": [
          {
            "name": "orden_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Orden Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrdenEdicion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "orders"
        ],
        "summary": "Archivar",
        "description": "Saca la orden del tablero. Es para la que se creo mal, no para la que se termino.\n\nLa que se termino se cierra moviendola a entregado, que es lo que le avisa al cliente.",
        "operationId": "archivar_orders__orden_id__delete",
        "parameters": [
          {
            "name": "orden_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Orden Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{orden_id}/status": {
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "Mover De Estado",
        "operationId": "mover_de_estado_orders__orden_id__status_post",
        "parameters": [
          {
            "name": "orden_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Orden Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CambioDeEstado"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/workshops/{taller_id}/orders": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Ordenes Del Taller",
        "description": "Las ordenes del taller, con los mismos filtros que ve el taller en su panel.\n\nMismos filtros a proposito: cuando el dueno dice \"no me aparece la orden de la\ncamioneta\", hay que poder mirar exactamente lo que el esta mirando.",
        "operationId": "ordenes_del_taller_admin_workshops__taller_id__orders_get",
        "parameters": [
          {
            "name": "taller_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Taller Id"
            }
          },
          {
            "name": "open",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Open"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100000,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/workshops/{taller_id}/orders/{orden_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Detalle De La Orden",
        "description": "La orden con su bitacora y todos sus avisos.\n\nEs la pantalla del post-mortem: quien la movio, cuando, y si el WhatsApp de cada\ncambio llego a salir o quedo esperando que alguien apretara enviar. El panel del\ntaller muestra solo el ultimo aviso -es lo que necesita para abrir wa.me-, y con eso\nno se puede contestar \"a mi cliente nunca le avisaron\".",
        "operationId": "detalle_de_la_orden_admin_workshops__taller_id__orders__orden_id__get",
        "parameters": [
          {
            "name": "taller_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Taller Id"
            }
          },
          {
            "name": "orden_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Orden Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CambioDeEstado": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "CambioDeEstado"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrdenEdicion": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "estimatedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimatedat"
          }
        },
        "type": "object",
        "title": "OrdenEdicion",
        "description": "Sin `status` a proposito: el estado se mueve por su propio endpoint.\n\nCambiar de estado le escribe al cliente. Si se pudiera hacer desde aca, corregir una\npalabra del titulo podria terminar mandando un WhatsApp sin que nadie lo pidiera."
      },
      "OrdenEntrada": {
        "properties": {
          "clientId": {
            "type": "string",
            "title": "Clientid"
          },
          "vehicleId": {
            "type": "string",
            "title": "Vehicleid"
          },
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "estimatedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimatedat"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "clientId",
          "vehicleId",
          "title"
        ],
        "title": "OrdenEntrada"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}